+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   ([\#5959](https://github.com/cometbft/cometbft/pull/5959))
 - `[mempool]` fix setRecheckFull/setDone race causing spurious ErrRecheckFull.
   ([\#5837](https://github.com/cometbft/cometbft/pull/5837))
+- `[mempool]` truncate proto field number to int32 in filter's ReadTag
+  ([\#5948](https://github.com/cometbft/cometbft/pull/5948))
 
 ### IMPROVEMENTS